refactor(ai-tools): share asset-field markdown helper; drop no-op assets frontmatter flag; test applyContentFormat
